Move TextureBufferStack and InvokeOnDisposal around.
