+++ osu.Framework/OS/BasicGameHost.cs
@@ -15,6 +15,7 @@ using OpenTK;
 using OpenTK.Graphics.OpenGL;
 using osu.Framework.Graphics.Performance;
 using osu.Framework.Statistics;
+using osu.Framework.Allocation;
 
 namespace osu.Framework.OS
 {
