+++ osu.Framework/Statistics/PerformanceMonitor.cs
@@ -1,6 +1,7 @@
 ﻿//Copyright (c) 2007-2016 ppy Pty Ltd <contact@ppy.sh>.
 //Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
+using osu.Framework.Allocation;
 using osu.Framework.Timing;
 using System;
 using System.Collections.Concurrent;
